Split parsing files
